requirements.txt
fastapi
uvicorn
python-dotenv
litellm
requests
pydantic
sqlalchemy
psycopg2-binary
redis
celery
qdrant-client
mcp
python-multipart
python-jose[cryptography]
passlib[bcrypt]
rich
typer
websockets
httpx
questionary
python-telegram-bot
python-telegram-bot
